+++ extra/release.py
@@ -34,7 +34,7 @@ VERSION_LOCS = [
         [
             (
                 r'__version__\s*=\s*u[\'"]([0-9\.]+)[\'"]',
-                "__version__ = u'{version}'",
+                "__version__ = '{version}'",
             )
         ],
     ),