+++ test/test_permissions.py
@@ -7,6 +7,7 @@ import unittest
 from unittest.mock import patch, Mock
 
 from test.helper import TestHelper
+from test._common import touch
 from beets.util import displayable_path
 from beetsplug.permissions import (
     check_permissions,
@@ -85,6 +86,26 @@ class PermissionsPluginTest(unittest.TestCase, TestHelper):
 
     def test_convert_perm_from_int(self):
         self.assertEqual(convert_perm(10), 8)
+
+    def test_permissions_on_set_art(self):
+        self.do_set_art(True)
+
+    @patch("os.chmod", Mock())
+    def test_failing_permissions_on_set_art(self):
+        self.do_set_art(False)
+
+    def do_set_art(self, expect_success):
+        if platform.system() == "Windows":
+            self.skipTest("permissions not available on Windows")
+        self.importer = self.create_importer()
+        self.importer.run()
+        album = self.lib.albums().get()
+        artpath = os.path.join(self.temp_dir, b"cover.jpg")
+        touch(artpath)
+        album.set_art(artpath)
+        self.assertEqual(
+            expect_success, check_permissions(album.artpath, 0o777)
+        )
 
 
 def suite():